Refactor and comment code
Before code refactoring, every time a new book object was created or updated, a displayBooks function was called. Every html card were removed, then recreated and added to the html list one by one. Not very good for memory.

Now, when a book object is created, it is simply added to the html list without the need to re-display every other books, and when a book object is modified, only the corresponding card element gets updated.